+++ src/connectors/tests/elastic.rs
@@ -15,6 +15,7 @@
 use std::time::{Duration, Instant};
 
 use super::ConnectorHarness;
+use crate::connectors::impls::elastic;
 use crate::errors::{Error, Result};
 use elasticsearch::{http::transport::Transport, Elasticsearch};
 use futures::TryFutureExt;
@@ -76,7 +77,7 @@ async fn connector_elastic() -> Result<()> {
             ]
         }
     });
-    let harness = ConnectorHarness::new(function_name!(), "elastic", &connector_config).await?;
+    let harness = ConnectorHarness::new(function_name!(), &elastic::Builder::default(), &connector_config).await?;
     let out = harness.out().expect("No pipe connected to port OUT");
     let err = harness.err().expect("No pipe connected to port ERR");
     let in_pipe = harness.get_pipe(IN).expect("No pipe connected to port IN");