+++ src/connectors/tests/s3/reader.rs
@@ -41,7 +41,7 @@ async fn connector_s3_no_connection() -> Result<()> {
         }
     });
 
-    let harness = ConnectorHarness::new(function_name!(), "s3_reader", &connector_yaml).await?;
+    let harness = ConnectorHarness::new(function_name!(), &s3::reader::Builder::default(), &connector_yaml).await?;
     assert!(harness.start().await.is_err());
     Ok(())
 }
@@ -71,7 +71,7 @@ async fn connector_s3_no_credentials() -> Result<()> {
         }
     });
 
-    let harness = ConnectorHarness::new(function_name!(), "s3_reader", &connector_yaml).await?;
+    let harness = ConnectorHarness::new(function_name!(), &s3::reader::Builder::default(), &connector_yaml).await?;
     assert!(harness.start().await.is_err());
 
     Ok(())
@@ -104,7 +104,7 @@ async fn connector_s3_no_region() -> Result<()> {
         }
     });
 
-    let harness = ConnectorHarness::new(function_name!(), "s3_reader", &connector_yaml).await?;
+    let harness = ConnectorHarness::new(function_name!(), &s3::reader::Builder::default(), &connector_yaml).await?;
     assert!(harness.start().await.is_err());
 
     Ok(())
@@ -134,7 +134,7 @@ async fn connector_s3_no_bucket() -> Result<()> {
             "endpoint": endpoint
         }
     });
-    let harness = ConnectorHarness::new(function_name!(), "s3_reader", &connector_yaml).await?;
+    let harness = ConnectorHarness::new(function_name!(), &s3::reader::Builder::default(), &connector_yaml).await?;
     assert!(harness.start().await.is_err());
 
     Ok(())
@@ -194,7 +194,7 @@ async fn connector_s3_reader() -> Result<()> {
 
     let harness = ConnectorHarness::new(
         function_name!(),
-        s3::reader::CONNECTOR_TYPE,
+        &s3::reader::Builder::default(),
         &connector_yaml,
     )
     .await?;